fix: eliminate SSRF by calling sheet counter logic directly instead of self-fetching
