+++ src/app/api/business/sheets/invoice-counter/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { isAdminRequest } from "@/shared/lib/auth";
-import { getSheetsClient, getSheetId } from "@/features/business/lib/google-sheets";
+import { getInvoiceCounter, setInvoiceCounter } from "@/features/business/lib/google-sheets";
 
 /**
  * GET /api/business/sheets/invoice-counter - Reads the next invoice number from the Google Sheet.
@@ -11,27 +11,9 @@ export async function GET(request: NextRequest): Promise<NextResponse> {
   if (!isAdminRequest(request)) {
     return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
   }
-
   try {
-    const sheets = getSheetsClient();
-    const spreadsheetId = getSheetId();
-
-    const res = await sheets.spreadsheets.values.batchGet({
-      spreadsheetId,
-      ranges: ["SETTINGS!B8", "SETTINGS!B11", "SETTINGS!B17"],
-    });
-
-    const ranges = res.data.valueRanges ?? [];
-    const prefix = (ranges[0]?.values?.[0]?.[0] as string | undefined) ?? "TTP";
-    const yearRaw = (ranges[1]?.values?.[0]?.[0] as string | undefined) ?? "";
-    const lastRaw = ranges[2]?.values?.[0]?.[0];
-
-    const yearCode = yearRaw.replace("-", "");
-    const lastNumber = lastRaw ? parseInt(String(lastRaw), 10) : 0;
-    const nextNumber = lastNumber + 1;
-    const nextFormatted = `${prefix}-${yearCode}-${String(nextNumber).padStart(4, "0")}`;
-
-    return NextResponse.json({ ok: true, lastNumber, nextNumber, yearCode, nextFormatted, prefix });
+    const data = await getInvoiceCounter();
+    return NextResponse.json({ ok: true, ...data });
   } catch (err) {
     console.error("[sheets/invoice-counter] GET failed:", err);
     return NextResponse.json({ error: "Sheet unavailable" }, { status: 503 });
@@ -47,7 +29,6 @@ export async function POST(request: NextRequest): Promise<NextResponse> {
   if (!isAdminRequest(request)) {
     return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
   }
-
   try {
     const body = await request.json();
     const newCount = Number(body.newCount);
@@ -57,17 +38,7 @@ export async function POST(request: NextRequest): Promise<NextResponse> {
         { status: 400 },
       );
     }
-
-    const sheets = getSheetsClient();
-    const spreadsheetId = getSheetId();
-
-    await sheets.spreadsheets.values.update({
-      spreadsheetId,
-      range: "SETTINGS!B17",
-      valueInputOption: "RAW",
-      requestBody: { values: [[newCount]] },
-    });
-
+    await setInvoiceCounter(newCount);
     return NextResponse.json({ ok: true, written: newCount });
   } catch (err) {
     console.error("[sheets/invoice-counter] POST failed:", err);
